requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.5
djangorestframework==3.13.1
gunicorn==20.1.0
pycodestyle==2.8.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
